Add basic elements for Subscription API.
You can now add and remove subscriptions. The notification part is not
finished, not even begun yet. But I hope to get my first notification
in soon.

First step would be sending a simple mail to myself to check if it is
working at all. Second is reacting to the notification and updating a
resource on behalf of a user.